Remove compiler warnings on Solaris 8. Can go into 2.2.x, but not necessary.

+++ Modules/_hotshot.c
@@ -1453,7 +1453,7 @@ get_version_string(void)
     char *buffer;
     int i = 0;
 
-    while (*rev && !isdigit(*rev))
+    while (*rev && !isdigit((int)*rev))
         ++rev;
     while (rev[i] != ' ' && rev[i] != '\0')
         ++i;
